Factor out BindingsSection from HoleCard
